Built Flask server
Built the intended Flask front-end withint server.py
Functionalities so far:
- The index page contains the last 20 Facebook-posts and the associated
data. Front-end needs some love but the data is there.
- Built a search engine: input a link to an article and the associated
Facebook post(s) will be found. Found some doubles in the database.
Integrity seems not to be compromised: data is the same as well.
- two templates: show_posts.html that extends into base.html. Base.html
includes the search-engine and all .css in static/style.css

To do:
- Flask: Improve feedback: list posts within an hour on the same week
day. List average performance of posts on same week day.
- Crawler: Probably need to rebuild the database/fbcrawler to include
weekday and hour. Better for performance and easier to code.
- Crawler: rebuild to refresh insights if posts is created within 24
hours from now.
- Rebuild to initialize from DB? Initializing from Facebook has its
merits. Need to think about this.
- Flask: improve css/styling.
- Flask: pageination: allow users to explore the database further.