Started daily view
- Started work on a new view, which contains the daily aggregate of
Facebook post performance.
- Currently /dagkoersen/ works with the current date only. A search
date search engine form is implemented, but does not yet connect to the
database: probably will have to handle it with a different functions,
which means that the base view will have to be updated as a different
action will be called when the user inputs the date.
 - Added some more error handling to the crawler and some suggestions
for a new data field.
- Implemented the data search form in server.py
- implement the code for the new view.
- Implemented the new view in show_day.html
- Had an issue with the title in show_posts: currently pasts the url
instead. In time, thinking of crawling the article and extracting the
title from the page.